+++ contracts/__init__.py
@@ -11,6 +11,7 @@ from .igraph_query import IGraphQuery
 from .snapshotable import ISnapshotable
 from .plugin import IPlugin
 from .embedding import IEmbedding
+from .desktop import IDesktop
 
 __all__ = [
     "IService",
@@ -22,4 +23,5 @@ __all__ = [
     "ISnapshotable",
     "IPlugin",
     "IEmbedding",
+    "IDesktop",
 ]